feat(landing/shop): split shop + gift-card sections, add shop gift-card CTA
- Landing order is now Creative Experiences -> Shopping -> Gift Cards, each
  separated by the wave divider (split the combined section into ShopPreview +
  GiftCardBanner)
- Shop page: add a gift-card CTA beneath the grid using the Coastal Creations
  gift-card visual (logo card) + "Buy a Gift Card" link
